Changes made during 1/21 meeting (added shooting autonomous and teleop improvements).
